client+mgmt: fix errcheck + math.MaxUint32 386-overflow on PR-D
- client/internal/peer/status_debounce_test.go: explicit _ = on
  rec.UpdatePeer*State/Meta calls so errcheck (golangci-lint) doesn't
  flag the discarded error returns. The tests deliberately ignore
  the return because the recorder is pre-populated with the peer.

- management/server/http/handlers/accounts/accounts_handler.go:
  cast math.MaxUint32 to uint64 when interpolating it into the error
  string. On 32-bit goreleaser builds (linux/386 etc.) the untyped
  constant 4294967295 cannot be passed where the printf %d formatter
  expects an int (overflows). uint64 covers all archs cleanly.

+++ management/server/http/handlers/accounts/accounts_handler.go
@@ -513,7 +513,7 @@ func validateUint32Timeout(name string, v int64) (uint32, error) {
 		return 0, fmt.Errorf("invalid %s: %d (must be >= 0)", name, v)
 	}
 	if v > int64(math.MaxUint32) {
-		return 0, fmt.Errorf("invalid %s: %d (exceeds %d)", name, v, math.MaxUint32)
+		return 0, fmt.Errorf("invalid %s: %d (exceeds %d)", name, v, uint64(math.MaxUint32))
 	}
 	return uint32(v), nil
 }
